+++ NEWS.md
@@ -1,3 +1,8 @@
+# cache 0.1.6
+
+ - `cache` now returns object invisibly for use in pipes.
+ - `uncache` now returns object invisibly for use in pipes.
+
 # cache 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.